market_fiyat_cekici/claude_code_package/market_fiyat_cekici/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.42.0
supabase>=2.4.0
python-dotenv>=1.0.0
lxml>=5.1.0
